Code at end of STL regional

+++ 2018 Competition Robot/src/org/usfirst/frc/team1732/robot/sensors/encoders/WPIEncoder.java
@@ -22,7 +22,7 @@ public class WPIEncoder extends EncoderBase {
 	}
 
 	@Override
-	public double getPulses() {
+	public int getPulses() {
 		return wpiEncoder.getRaw();
 	}
